Assistant checkpoint: Fix practice projects positioning and styling
Assistant generated file changes:
- templates/tutorials.html: Fix practice projects positioning and remove duplicate styles

---

User prompt:

still do not change or remove anything in the current tutorial.html page i just want you look at this code i have sent you from github and check the styling of the practice project games, save it in the current tutorials.html so that the practice projects can display well on the website. remember to remove dublicate code in the current tutorials.html page. nothing has happened yet why is the positioning having issues because on the website the practictice projects are not positioned and styled well what could be the issue did you go through this code the way i instructed you above. go through all your code to find out why the practice projects are having issues. again do not remove or change anything in the current tutorials.html page just remove the duplicate and redandant code that we are not using in our website