Assistant checkpoint: Update Fun Interactive Games section and fix age calculator
Assistant generated file changes:
- templates/tutorials.html: Update Fun Interactive Games section with informative text, Fix Age Detective game calculation

---

User prompt:

in the tutorials.html Fun Interactive Games include a text that informs the user that they need to try out differrnt answers to see how the conditional statements if elif and else works. also fix the issue with the age detective game because it seems it is not calculating the age well according to the year the user enters. remember to remove duplicate and redundant code that we are not using in the code.