Assistant checkpoint: Add informative message about future game improvements
Assistant generated file changes:
- templates/tutorials.html: Add informative message about game improvements

---

User prompt:

in the tutorials.html in the Interactive Python Games, below the games, inform the user/learner that the games are a great improvement in their coding skills and am proud of them and that in the next tutorials we will learn how to write code that restrict inputs that are out of range or invalid like in the age explorer game when you put negative years like -1995 it will still give the answer which should not be the case. also in the little star game if you put something like 89 years you are still getting the same responce that someone who is 15 years which should also not be the case. for now just play the games as they are and think of how to write a code that will give valid information to our user. Now on your own style this text and position it proffessionally the best way you can so that when the users read it they will understand. use font colour and style that you think is the best.