Assistant checkpoint: Add game preview images with professional styling
Assistant generated file changes:
- static/nickname_game.png: Save Sweet Nicknames game image
- static/age_explorer.png: Save Age Explorer game image
- static/little_star.png: Save Little Star game image
- templates/tutorials_part2.html: Add game preview images

---

User prompt:

in the /tutorials_part2 directly above the Want to see how these games are different? section put these image, save them in the static file so that flask can be able to use them, give them the heading 💝 Sweet Nicknames 🎂 Age Explorer ⭐ Little Star, respectively so that the user is able to know which image belongs to which game.style and position the images profesionally use the font or style or the colour that is best.