Assistant checkpoint: Add game preview images and styling to tutorials_part2.html
Assistant generated file changes:
- copy_images.py: Add image copy script
- templates/tutorials_part2.html: Add game previews
Add game previews

---

User prompt:

i can see the images i just sent you in the attached_assets but i cannot see them in the static file. please make copies of these images and store them in the static files so that flask is able to access and display them so that we are able to carry out these in the /tutorials_part2 directly above the Want to see how these games are different? section put these image, save them in the static file so that flask can be able to use them, give them the heading 💝 Sweet Nicknames 🎂 Age Explorer ⭐ Little Star, respectively so that the user is able to know which image belongs to which game.style and position the images profesionally use the font or style or the colour that is best.