Assistant checkpoint: Update game styles and add image popup functionality
Assistant generated file changes:
- templates/tutorials_part2.html: Add modal HTML and JavaScript

---

User prompt:

in the tutorials_part2 make the games colourful with attractive colours for each game that is appealing to the users. then at this time the 3 images are displaying but the user cannot properly see the code in them,  make it so that the  3 images pop up and become bigger  when clicked so that the user is able to see the code well that is being displayed by the images.